feat: single SOT for themes + basic custom support
