+++ client/package.json
@@ -14,6 +14,7 @@
     "@react-router/node": "^7.5.3",
     "@react-router/serve": "^7.5.3",
     "@tanstack/react-query": "^5.80.6",
+    "@vanilla-extract/css": "^1.17.4",
     "color.js": "^1.2.0",
     "isbot": "^5.1.27",
     "lucide-react": "^0.513.0",
@@ -27,6 +28,7 @@
     "@types/node": "^20",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
+    "@vanilla-extract/vite-plugin": "^5.0.6",
     "tailwindcss": "^4.1.4",
     "typescript": "^5.8.3",
     "vite": "^6.3.3",
